TJ: create meresco-normalize 3.0-beta version 0
It is based on the Delft-Tilburg project where normalizations were created for year and names. 
It is also based on the Discover project from Delft where normalizations for languages were made based on the iso639 standard.
The code from the normalizations of these projects have been donated to the Meresco project and are now part of Meresco Normalizations.

git-svn-id: file:///home/thijs/development/meresco2git/meresco/meresco-normalize/workingsets/3.0-beta/version_0@2153 83e7a582-4515-4195-8f39-222cbda6297e